force to int instead of float

+++ src/moonchild/engine.cpp
@@ -33,18 +33,18 @@ static float read_number_value(moon_value * value) {
   return data;
 }
 
-static float read_constant(moon_prototype * prototype, uint16_t index) {
+static int32_t read_constant(moon_prototype * prototype, uint16_t index) {
   moon_value value;
-  float data = 0;
+  int32_t data = 0;
 
   progmem_cpy(&value, prototype->constants_addr, sizeof(moon_value), sizeof(moon_value) * index);
 
   switch(value.type) {
     case LUA_NUMBER:
-      data = (float) read_number_value(&value);
+      data = (int32_t) read_number_value(&value);
       break;
     case LUA_INT:
-      data = (float) read_int_value(&value);
+      data = read_int_value(&value);
       break;
     default:
       break;
@@ -91,5 +91,5 @@ void moon_run(PGMEM_ADDRESS prototype_addr, char * result) {
     run_instruction(&closure, index);
   }
 
-  sprintf(result, "reg0 = %f", closure.registers[0]);
+  sprintf(result, "reg0 = %d", closure.registers[0]);
 }
